Rename package org.apache.calcite.slt to net.hydromatic.sqllogictest
Capitalize acronyms in class names; for example, JDBCExecutor
becomes JdbcExecutor.
