+++ src/main/java/net/hydromatic/sqllogictest/SqlTestQueryOutputDescription.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.apache.calcite.slt;
+package net.hydromatic.sqllogictest;
 
 import java.util.ArrayList;
 import java.util.List;
